Trunk, MTXml: changed entity lookup from linear string array search to a perfect hash map, generated by GNU gperf. Also added all the standard Unicode entities.
git-svn-id: svn://localhost/public/MoSync/trunk@909 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ testPrograms/mtxMoblet/mtxMoblet.cpp
@@ -230,7 +230,7 @@ private:
 				dumpStats();
 		}	else {	//(mTortureTest)
 			mPtr[result] = 0;
-			LOG("feed %i\n", (mPtr - mBuffer) + result);
+			LOG("feed %li\n", (mPtr - mBuffer) + result);
 			mPtr = mBuffer;
 			int parseStartTime = maGetMilliSecondCount();
 			if(mProcessing)
@@ -300,7 +300,7 @@ private:
 	virtual void mtxTagData(const char* data, int len) {
 		m.tagData++;
 		LOG("d %i: \"%s\"\n", len, data);
-		if(len > BUFLEN || len <= 0 || strlen(data) != len) {
+		if(len > BUFLEN || len <= 0 || (int)strlen(data) != len) {
 			ERROR("d %i: \"%s\"\n", len, data);
 			mError = true;
 			return;
